+++ ESFramework/ESMacros.h
@@ -9,7 +9,6 @@
 #ifndef ESFramework_ESMacros_h
 #define ESFramework_ESMacros_h
 
-#import <mach/mach_time.h>
 #import "ESHelpers.h"
 #import "NSInvocation+ESHelper.h"
 
@@ -50,24 +49,6 @@
  */
 #define ESDefineAssociatedObjectKey(name) static const void * name##Key = &name##Key;
 
-/**
- * Log the execution time.
- */
-#if DEBUG
-#define ES_STOPWATCH_BEGIN(stopwatch_begin_var) uint64_t stopwatch_begin_var = mach_absolute_time();
-#define ES_STOPWATCH_END(stopwatch_begin_var) { \
-        uint64_t end = mach_absolute_time(); \
-        mach_timebase_info_data_t timebaseInfo; \
-        (void)mach_timebase_info(&timebaseInfo); \
-        uint64_t elapsedNano = (end - stopwatch_begin_var) * timebaseInfo.numer / timebaseInfo.denom; \
-        double_t elapsedMillisecond = (double_t)elapsedNano / 1000000.0; \
-        printf("‼️STOPWATCH‼️ [%s:%d] %s %fms\n", [NSString stringWithUTF8String:__FILE__].lastPathComponent.UTF8String, __LINE__, __PRETTY_FUNCTION__, elapsedMillisecond); \
-}
-#else
-#define ES_STOPWATCH_BEGIN(stopwatch_begin_var)
-#define ES_STOPWATCH_END(stopwatch_begin_var)
-#endif
-
 #pragma mark - Constants
 
 #define ES_MINUTE   (60)
@@ -75,7 +56,7 @@
 #define ES_DAY      (86400)
 #define ES_5_DAYS   (432000)
 #define ES_WEEK     (604800)
-#define ES_MONTH    (2635200) /* 30.5 days */
+#define ES_MONTH    (2635200)  /* 30.5 days */
 #define ES_YEAR     (31536000) /* 365 days */
 
 #endif /* ESFramework_ESMacros_h */