+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     + `NSDateFormatter (ESAdditions)` category
     + `-[NSUserDefaults setObject:forKeyPath:]`
 - Renamed:
+    + `ES_STOPWATCH_BEGIN` `ES_STOPWATCH_END` => `ESBenchmark()`
     + `ESDispatchOnMainThreadAsynchrony` => `es_dispatch_async_main`
     + `ESDispatchOnMainThreadSynchrony` => `es_dispatch_sync_main`
     + `ESDispatchOnGlobalQueue` => `es_dispatch_async_global_queue`